refactor: AbstractRestaurantTimeFactory로 인터페이스 대체
